# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tunploy API",
    "version": "1.2",
    "description": "Manage devices and servers on your Tunploy VPN from your own programs, and hear about what happens to them.\n\nSend an API key from Settings → API keys as `Authorization: Bearer tp_…`. Each key only does what its scopes allow. Requests are limited to 10 a second per key, with bursts up to 60. Keep keys on your server, never in an app.\n\nErrors always look like `{\"error\": {\"code\": \"…\", \"message\": \"…\", \"fields\": {…}}}`; `fields` names the inputs a validation error is about."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/devices/{id}": {
      "get": {
        "operationId": "getDevice",
        "summary": "Get a device",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:read",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the `devices:read` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          }
        ]
      },
      "patch": {
        "operationId": "updateDevice",
        "summary": "Change a device",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Invalid"
          },
          "502": {
            "description": "`apply_failed`: saved, but the running tunnel could not be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Only the fields sent change. To put the device on another server use `POST /devices/{id}/move`; the public key cannot change.\n\nNeeds the `devices:write` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 64,
                    "description": "Unique per server. Made up from external_id when left out."
                  },
                  "external_id": {
                    "type": "string",
                    "maxLength": 255
                  },
                  "metadata": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "additionalProperties": true,
                    "description": "Any JSON object up to 4 KB; null clears it."
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "data_limit": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "limit_period": {
                    "$ref": "#/components/schemas/LimitPeriod"
                  },
                  "expires_at": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time",
                    "description": "null clears it."
                  },
                  "speed_limit": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000000,
                    "description": "kbit/s (1000 bits a second), download and upload each; 0 means no limit."
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteDevice",
        "summary": "Delete a device",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the `devices:write` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          }
        ]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key: `tp_…`."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "`unauthorized`: no key, or it is unknown, revoked or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "unauthorized",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "`missing_scope`: the key lacks the scope this needs.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "missing_scope",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "`not_found`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "not_found",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "Invalid": {
        "description": "`validation_failed`: `fields` says what is wrong with which input.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "validation_failed",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "`rate_limited`: wait for Retry-After seconds.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "rate_limited",
                    "message": "…"
                  }
                }
              }
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Traffic": {
        "type": "object",
        "required": [
          "rx_bytes",
          "tx_bytes",
          "total_bytes"
        ],
        "description": "Seen from the server: rx is what the device uploaded, tx what it downloaded.",
        "properties": {
          "rx_bytes": {
            "type": "integer"
          },
          "tx_bytes": {
            "type": "integer"
          },
          "total_bytes": {
            "type": "integer"
          }
        }
      },
      "DeviceStatus": {
        "type": "string",
        "enum": [
          "active",
          "disabled",
          "expired",
          "limit_reached"
        ]
      },
      "LimitPeriod": {
        "type": "string",
        "enum": [
          "monthly",
          "total"
        ],
        "description": "monthly counts from the first of each month in the panel's time zone; total counts until a usage reset."
      },
      "Device": {
        "type": "object",
        "required": [
          "id",
          "server_id",
          "name",
          "external_id",
          "metadata",
          "address",
          "public_key",
          "client_key",
          "enabled",
          "status",
          "online",
          "last_handshake",
          "data_limit",
          "limit_period",
          "period_usage",
          "usage_reset_at",
          "expires_at",
          "speed_limit",
          "month_usage",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "server_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "external_id": {
            "type": "string",
            "description": "Your own ID for the device's owner. Not unique."
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "address": {
            "type": "string",
            "description": "The device's address inside the tunnel.",
            "examples": [
              "10.8.0.2"
            ]
          },
          "public_key": {
            "type": "string"
          },
          "client_key": {
            "type": "boolean",
            "description": "The client made the key pair; the panel never saw the private key."
          },
          "enabled": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          },
          "online": {
            "type": "boolean",
            "description": "As of the panel's last look, at most ten seconds old."
          },
          "last_handshake": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "data_limit": {
            "type": "integer",
            "minimum": 0,
            "description": "Bytes per limit period, both directions; 0 means no limit."
          },
          "limit_period": {
            "$ref": "#/components/schemas/LimitPeriod"
          },
          "period_usage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Traffic"
              }
            ],
            "description": "What counts toward data_limit: this month, or since the last usage reset."
          },
          "usage_reset_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "The last usage reset, if any."
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Access ends at this instant."
          },
          "speed_limit": {
            "type": "integer",
            "minimum": 0,
            "description": "kbit/s (1000 bits a second), download and upload each; 0 means no limit."
          },
          "month_usage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Traffic"
              }
            ],
            "description": "The calendar month, whatever the limit period."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "config": {
            "type": "string",
            "description": "Only in the reply to a create: the WireGuard config. Holds the private key unless public_key was sent."
          }
        }
      },
...
    }
  }
}